diff_util: move unified diff utils to `jj-lib`
